Improve demo and fix inner outState
Improve the demo script and README.md file for it,
explaining in detail the deletion requirement which
is the more tricky.

Also, fixes an important detail: the inner event out-state
was contaminated with the foreign method state. This could
be confusing and also alterates without real desire, the
state for the current event (DELETE).

Change-Id: I5ae45a7e8ab79080495601f636e9b23a795ceb72