Implement new transformation source: event
Added source event documentation to README.md file.
dded component tests and tools test.

Also,

Improved provision schema patterns for source and target:
fixed some weak validation (for example, 'response.body.whatever'
was wrongly accepted).

Refactored the server-data results. When all is requested,
method and uri appears as keys in a superset (array). For
every key, we have the requests array. But when we request
a key (method+uri), we will obtain the requests array. If
we also specify a requestNumber (not 0 which means everything),
then the single object is retrieved (not an array). If 0 is
provided, an array will be retrieved even if one only event
exists.

Demo slighly improved.

Updated README.md file.