Respect history for unprovisioned events
Even if history configuration is set to be discarded,
those unknown requests (with no provisions matching reception
criteria) will be stored anyway with a requests history for the
method/uri unhandled by the mock configuration.

In case that general server-data is discarded, the history
is also discarded even for unprovisioned events.