Add prometheus-cpp library to the builder image
Also, we simplify the builder script to remove
the default value as argument in read function
as we could know the name of that variable and
get the value with eval command.

Also, help is improved to show '--auto' in
non-cached execution example, as it is probably
more common and could be copy/pasted easily.