Fix 'matches source out of scope' bug
This commit fixes an important issue regarding the
use of std::smatch as container. This object refers
iterators to the passed source string, so when this
string is destroyed or go out of scope, such iterators
are pointing to garbage. This was causing problems in
some transformations based on RegexCapture filter onto
a target variable.

Also,

- improve tracing:
  - one-line traces
  - add some debug traces and improve current ones

- move base images to registry ghcr.io as well as
  CI workflow implementation.