Fix purge state storage
When state was 'purge' and purge execution was disabled,
this context was not being stored.

Also,

- we will omit unprovisioned register state fields
(inState, outState) which are empty. To avoid ambiguity,
we convert the user provision empty states into 'initial',
in this way empty values are only provided by unprovision
situation anbd will be properly hiden.

- Fix 'RESPONSE TO SENT' into 'RESPONSE TO SEND'

- Add metrics counter for purge operation (successful and failed
labels).

- Add 'sequence()' helper function to easily access the server
data register for a specific serverSequence value.