Add examples to helper functions
Also, schema, matching and provision allow now to
pass configuration file to process.
Note that schema admit h2agent schema structure,
which is not a json schema but a structure with
the identifier and the inner shema.

Examples output can be sourced or studied and
is not destructive with eventual agent configuration.