Refactor names to prepare client feature
This is a key commit to help to distinguish which
is related to mock server feature before starting
to implement the client feature for h2agent [0].

H2agent is not only a mock server. The intention is
to use it within an agents-oriented architecture,
covering both server and client control plane for
HTTP2 protocol. So, here we will name every related
to current implementation (server plane) adding
the word 'server' to things like provision (it is
a server-provision), matching or data (although it
will be also client-data). Also 'globals' is renamed
to 'global-variable'.

This affects to classes, files, variables, etc.

This commit also changes the relative URI for global
variables, which was hold over server-data but it
is more generic, and will be now an independent
'/admin/v1/global-variable' path. Also, delete
operation has been separated from server-data
deletion.

README.md is updated with all the things renamed and
also the command-line help updates, helpers and
benchmarking script output.

Data configuration is renamed to change requests by
key, because this will be shared with client features
and the key will the target endpoint instead of method/uri,
and requests is not a good name to describe the history set.

Also,

- New bind_address parameter for commandline, falling back
  to previous hardcoded default: 0.0.0.0
- Server mode can be disabled by command-line: port = -1
  and startup output will be reduced to only general and
  future client properties.
- README.md improved to have a summary at management interface
  describing the sections: general/mock server

[0] https://github.com/testillano/h2agent/issues/13