Improve histogram boundaries help
When providing histogram boundaries, it is good to know that
scientific notation is allowed. We will improve the command
line help to explain this. The input interpreter is also
improved to sort the values regardless if the user was wrong
when providing them.

Also, message size boundaries apply to both Rx and Tx of traffic,
(http2comm library does not allow to distinguish boundaries for
reception and transmision, the same are used for both).