Fix request body string interpretation
Fixed a problem [0] getting request body string (not a json)
into response body as string. This may be interpreted as raw
string, not as a json.

Also, we implement a new feature to merge response body template
with possible transformations which move objects into final
response body object. It is better to merge because you could
remove the things you don't want, but it is not possible to
add a response template once it is overwritten [1].

[0] https://github.com/testillano/h2agent/issues/16
[1] https://github.com/testillano/h2agent/issues/17