Improve manual testing tool
Instead of implement a transform helper, we improve the current
utility to surf provisions and requests in order to test the
h2agent process. A transform helper would repeat an important
part of the agent source code, just changing http2 interfaces
by command-line or file-system based inputs. But this is overkilling
for the purpose.

See [0].

[0] https://github.com/testillano/h2agent/issues/28