Refactor matching configuration
Query parameters separator (normally &, but in rare/old systems ; could be used),
is coupled to the QP filter. So you have SortAmpersand, SortSemicolon, PassBy
and Ignore, but you don't have PassBy and Ignore for Semicolon (it is said as
assumed limitation, as semicolon separator is rare).

This commit, decouples them to better combine and make a more flexible
configuration.

Also, README file is improved:
- Better explanation for state machine in provisions.
- Better explanation for event sources.

Implements [0]

[0] https://github.com/testillano/h2agent/issues/41