Deprecate PriorityMatchingRegex
Implements [0]: PriorityMatchingRegex algorithm is renamed to
RegexMatching, which is more intuitive and is named closely to
FullMatching and FullMatchingRegexReplace.

Anyway, this algorithm name is still accepted although a
warning trace of deprecation note is traced to encourage
the user to migrate the matching algorithm.

The behaviour is exactly the same, and README.md file has been
updated also to better explain the three classification algorithms.

UT & CT have been updated to use the new algorithm name.
The method findWithPriorityMatchingRegex() is renamed to findRegexMatching().

Also, other minor changes have been done:
- Set -Ofast optimization flag to Release compilation.
- Helpers remove spaces in header content type expressions because
  curl command will show something without quotes and removing
  spaces we can copy/paste the printout and it would work on
  -H parameter.

[0] https://github.com/testillano/h2agent/issues/42