Cache static response bodies
This is an important improvement for those provisions
which respond static responses not affected by dynamic
transformations.

Currently we are invoking nlohmann::json dump() for every
transaction. Now, we will respond the already stored
cached string representation for that unvariable response
body.

The logic to know if transformations will be needed or not
over that response object is already present, so this
commit is an easy modifications with a great performance
improvement, specially for large reponse bodies, under the
conditions commented above.

For example for a 1.5k - sized response body, we double the
speed of the system:

H2load results for 1MM requests and m=20, request body = 1.5k
and h2agent provision with 1.5k static response body:

Before              Now improved
------------------- ------------------------
100k req/s (196 us) 189k req/s (103 us)

Issue: https://github.com/testillano/h2agent/issues/49