Convert BodyData into composed class
BodyData was created as typedef to old std::string type
received by the server and processed by the mock logic.
Now, we will transform the BodyData into class type which
contains an std::string and use move semantics to keep
performance not affected.

This commit represents the first step to improve the
abstraction of reception processing to manage different
content types.