Rename BodyData to DataPart
In order to prepare the source to support multipart
http bodies, we will rename the data vault to host the
request or responses data part.
This is because multipart content-type cannot be nested
and it would be strange to name BodyData to any of its
data parts. The whole request body could be named with
the old name, but DataPart is also valid. We will rename
the instances of request bodies to requestBodyDataPart
to improve the naming and clarify the intention.