Add new asAsciiString function
New function and data part getter added.
This function displays characters which are printable
and uses dots for non-printable  ones. It is useful for
traces, although 'as hex string' representation could be
also suitable. But that will be used on data storage
registers becuase it is reversible.

Added UT for functions.cpp and DataPart class.

also, decoded indicator for lazy decode is added to avoid
repeating this costly operation.