Refactor request body data decoding
We adapt the main reception logic to the new
DataPart class to process the request body, simplifying
provision interfaces.

Content-type is lazy decoded also in data part decode
procedure, to speed up the flow, because formerly it
was done in any case but it is only needed if decoding
is finally required.

With this change, everything is ready to implement
multipart processing.

We also protect the case of receiving binary data and trying
to transform over json response target (exception catched
with an error log).

Unit-tests are adapted to the new implementation.