Add 'EqualTo' filter type
Implements new feature [0].

A filter comparison is useful to match
two strings. This could be used to check the
request body content or any other source.

The advantage against RegexCapture is that
special symbols would require complex
regular expressions to do the same thing.

This new filter is normally used with ConditionVar
to transfer data conditionally with a target variable
built with 'EqualTo' transformation. Check README for
details.

New UT and CT are added for this filter.

Also,
- fix st/start.sh bug regarding final report which didn't consider
  H2AGENT__FILE_MANAGER_ENABLE_READ_CACHE_CONFIGURATION variable.
  Also, 'st/last.sh' is renamed to 'st/repeat.sh' allowing its use
  with any other report, not only the last one (default).
- Added skip logs on transformations
- Remove trailing decimal zeroes on float storage in order to
  emulate stringstream formatting when integer is detected for
  a double type. This make possible to have 1/0 for conditional
  math expressions instead of 1.000000/0.000000 due to the way
  type converter is manipulating the information internally (to_string).

[0] https://github.com/testillano/h2agent/issues/65