Make max worker threads configurable
Before this commit, the number of worker threads was a
fixed value which needed to be (probably experimentally)
tunned depending on the hardware concurrency (number of
host CPUs) and the importance of I/O times spent by
working threads during mock processing.

Now, there is a posibility to configure a maximum [0]
number of threads to which the used threads could grow to.
So, when needed, the number of worker threads will be
increased until no more are consumed. The user must be
sure that the maximum number does not compromish the
"context swtich" penalization.

README.md updated with this new feature and program
outputs.

Also,
 * Minor improvements in README.md file.
 * tools/manual is renamed to tools/play-h2agent.
 * skip IO service deletion on program exit.
 * Show hardware concurrency in help output.
   This can be useful to know the number to pass
   on '-w' and the maximum number (+50% ?) for the
   maximum number of worker threads.
 * Improve benchmarking script to add prepend
   form on variables load.
 * Add Arash Partow help URL on program utility help.

[0] https://github.com/testillano/h2agent/issues/67