Fix request.uri source to get original one
Provision transformation input was getting request
classification uri instead of original one, when
we had query parameters.

So we will improve traces to distinguish classification
and original URI, and hence provide the original one
for `request.uri` source.

Also, we improve README.md to explain the difference
between `RegexCapture` and `RegexReplace` regarding
the output result when matching is not fulfilled.
The regex replace will fallback to original unmatched
string, which is useful for traffic classification
related to incoming URIs not expected. The other one
is useful to bild condition variables which are those
with empty/undefined value for false boolean meaning.

Change-Id: Icc5394a6fd049e5419c433013ff3876e58ad7d55
Fix: https://github.com/testillano/h2agent/issues/72