refactor play-h2agent stuff
Manual interactive utility (play-h2agent) is confunsing and
complex to be a quick refrence for the h2agent service.

So we will build a better approach with tiny examples:

examples/
├── Evolution
├── ExtractTrailingNumberInURI
├── RandomNumericResponse
├── TheLastOfUs
└── a+bx

They could grow in future with interesting situations, and
the idea is to show a guided tour through each problem,
present the configuration and launch the request to get
the results easily.

It is a good complement (and less effort needed) to demo and
kata stuff.