Restore request.uri source to be normalized URI
Briefly, some of recent releases considered that the
source `request.uri` was the original URI (having or
not query parameters).

But it is better to restore previous behavior to
normalize such URI being more predictable.

This is, again, non-backward compatible, but
fortunately had no important impact as normally,
regex transformations are applied over URIs
with no query parameters and when done, those
parameters are normally sorted.

Also, README file is improved regarding packaged
utitilies (matching-helper and arashpartow-helper),
in order to run them with docker.