Rename requestNumber to eventNumber
Improvement to be more abstract, as client features are being added.
Events are requests towards server but also include response information,
 so it is a whole event.
Same for client: events are requests sent but also possible responses
received. All the information is an event.

Indeed, accessing the event you could access i.e. requestBody, but also
responseBody, so requestNumer is not a good name for the query parameter.

Fixes: https://github.com/testillano/h2agent/issues/78

Jira: