Rename event source to serverEvent
As future proof we will be able to access both server
and client events from server and client provisions,
so if in a specific provision (for example for server)
we want to retrieve data from client endpoint, we should
use the source clientEvent. At the moment that part is
not implemented, but we will adapt the current event
source (for server data) to be named: serverEvent to
allow distinguish them when coth are available.

Implements: https://github.com/testillano/h2agent/issues/80

Jira: