Fix found state for unprovisioned events
When an event is unprovisioned, register will miss
the state information (indeed this field is not
shown when server data is inspected).

But when finding last registered state, the event
is found, so state returned is empty but should be
"initial". We must add the condition to check if the
state is empty and then set the "initial" default value
considering this event as if it would be ignored (never
happen).

Also, we add a warning for the case when we change the
matching algorithm having already provision items
configured. This is because some algorithm stores the
information in different manner and could work in
different way than expected. So, the normal use implies
to first configure matching algorithm and then make the
provisions. Or just remove the provisions after matching
configuration update.

Fixed: https://github.com/testillano/h2agent/issues/83