Reverse condition variable
Implement the negated condition for variables.
This is done prefixing the variable name with
the exclamation mark '!'.

Also, we improve the provision schema to deny
empty variable names, and in the case of the new
feature, allow in general any non empty string
except single "!". So "!!" is valid, which means
negation of the variable called "!", although seems
to be very weird.

Implement: https://github.com/testillano/h2agent/issues/82

Jira: