Refactor event source
This is a NON-BACKWARD compatible change,
but it worth the improvement to define
event sources in a more compressed way.

The format is fully documented in the
updated README.md file, in summary, we
will use a query parameters format with
those keys:
requestMethod, requestUri (may be encoded),
eventNumber and eventPath (optional).

Implements: https://github.com/testillano/h2agent/issues/80

Jira: