Improve README.md for filters ConditionVar/EqualTo
Improved documentation with examples to better
understand these important filters.
Traces have been improved for them to know if
the condition is fulfilled or not (also for negated
condition variable).

Also, there is a minor non-backward compatible change for
'EqualTo' filter. Documentation said that this condition
is used to transfer source into target and at the same time
says that it takes the value "yes" when it is true. This is
incoherent. We consider more useful to transfer the source,
like condition variable does. So, we will fix the documentation
to be right that and modify source as well as CT to cover the
use case.

This minor fix may be taken into account when the filter is used.

Fixed: https://github.com/testillano/h2agent/issues/88