Simplify proxy mode configuration
There is no need to keep emable proxy flag.
Just defining the H2AGENT_HTTP1_PORT is enough to signal
that nghttpx proxy is going to be launched.
Then, if H2AGENT_HTTP2_PORT is provided (defaults to 8000),
must be aligned with -p traffic port parameter, and if -p
is provided, the same.