Implement dynamic reponses delay
New http2comm version to v2.1.11 provides a virtual method
of the Http2Server, to implement the way to extract the
response delay for an specific reception.

In h2agent, this will be based on global variables, creating
a new one with the name: ResponseDelayTimerUS.ReceptionId.<recvseq>
Where recvseq is the received sequence, a unique monotonically
increasing sequence value to identify the server reception id.

Also,

- Update README.
- Install socat within project image.
- Improved udp-server-h2client utility to dump,
  optionally, an UDP datagram on output socket
  when responses are received. This way, we can
  signal this event to trigger other instances
  which will launch another request, to do a chain
  of requests, or trigger an administrative request
  which could remove the global variable to stop
  the dynamic response delay on synchronous request
  to respond it immediately.